refactor(macos): extract item actions & sharing from AppModel into AppModel+ItemActions.swift
Continues the AppModel god-object split. Relocates the context-menu actions and
share links for albums, artists, and tracks — play/shuffle/playNext/addToQueue/
instant-mix radio, favorite & follow toggles, multi-select favorite/download/
mark-played, add/remove-from-playlist, navigation jumps, and Jellyfin web-link
sharing — into a new `extension AppModel` file. The SongRadioRoute enum moves
with it.

Pure relocation: the 372 moved lines are byte-identical to what was removed.
AppModel.swift drops 6,518 -> 6,146 lines. The one stored property in the
cluster (trackInfoSubject, the track-info sheet anchor) stays on the main class
under a new MARK; stored properties can't live in an extension.

Two access promotions, both shared action primitives the moved methods route
through: playInstantMix (private -> internal; 11 call sites across albums,
artists, tracks, genre, now-playing) and setPlayed (private -> internal; also
called by album mark-played). loadTracks(forArtist:) is self-contained and
stays private in the new file.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
